Test fix

+++ src/models/index.js
@@ -1,15 +1,15 @@
-import { Sequelize } from 'sequelize';
+import { Sequelize, DataTypes } from 'sequelize';
 
 import UserModel from './user';
 import EntryModel from './entry';
 import env from '../utils/env';
 
-const sequelize = new Sequelize(env.databaseURL, { ssl: true, dialect: 'postgres' });
+const sequelize = new Sequelize(env.databaseURL, { ssl: true, dialect: 'postgres', logging: false });
 // pass your sequelize config here
 
 const models = { user: UserModel, entry: EntryModel };
 
-Object.values(models).forEach((model) => model.init(sequelize));
+Object.values(models).forEach((model) => model.init(sequelize, DataTypes));
 
 // Run `.associate` if it exists,
 // ie create relationships in the ORM
@@ -19,8 +19,9 @@ Object.values(models)
 
 (async () => {
   await sequelize.authenticate();
-  await sequelize.sync({ force: true, match: /dev$/ });
-  // no sequelize.sync(); use umzug migrations after writing models
+  if (process.env.NODE_ENV !== 'test' && process.env.NODE_ENV !== 'production') {
+    await sequelize.sync({ force: true, match: /dev$/ });
+  }// no sequelize.sync(); use umzug migrations after writing models
 })();
 
 export default {
